#198 Move send-inn button in a new component
